# openapi.yaml
openapi: 3.0.1
info:
  title: Travelsite API
  description: This is the Travelsite API
  version: 1.0.0
externalDocs:
  description: Github
  url: https://github.com/Kaiserlucas/travel-site-backend
servers:
  - url: https://webdevelopment-travelsite.herokuapp.com/
tags:
  - name: trips
    description: Everything about your trips

paths:
  /trips:
    get:
      tags:
        - trips
      summary: Get all trips
      responses:
        "200":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedTrip'
    post:
      tags:
        - trips
      summary: Add a new trip
      requestBody:
        description: Trip object that needs to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trip'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedTrip'

  /trips/{uuid}:
    delete:
      tags:
        - trips
      summary: Deletes a trip by uuid
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          description: uuid of the entry to delete
      responses:
        "204":
          description: successful operation

  /login:
    post:
      tags:
        - login
      summary: Log in and get cookie for authentication
      responses:
        "200":
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'

  /signup:
    post:
      tags:
        - signup
      summary: Sign up and create a new account
      responses:
        "200":
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'

components:
  schemas:
    Trip:
      type: object
      properties:
        name:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    SavedTrip:
      allOf:
        - $ref: '#/components/schemas/Trip'
        - type: object
          required:
            - uuid
          properties:
            uuid:
              type: string
              format: uuid
    Login:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string